requirements.txt
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9
openai>=1.0.0
